Refactored the WebSocket logic
