+++ src/features/mode/play/dialog/AcceptDrawDialog.js
@@ -2,7 +2,6 @@ import { useDispatch, useSelector } from 'react-redux';
 import { Button, Dialog, DialogActions, DialogContent, DialogTitle } from '@mui/material';
 import * as actionConst from 'features/mode/actionConst';
 import * as playMode from 'features/mode/playModeSlice';
-import Ws from 'socket/Ws';
 
 const AcceptDrawDialog = () => {
   const state = useSelector(state => state.playMode);
@@ -11,13 +10,19 @@ const AcceptDrawDialog = () => {
 
   const handleDrawAccept = (event) => {
     event.preventDefault();
-    Ws.draw(actionConst.ACCEPT);
+    dispatch({
+      type: 'socket/draw',
+      payload: actionConst.ACCEPT,
+    });
     dispatch(playMode.acceptDrawDialog({ open: false }));
   };
 
   const handleDrawDecline = (event) => {
     event.preventDefault();
-    Ws.draw(actionConst.DECLINE);
+    dispatch({
+      type: 'socket/draw',
+      payload: actionConst.DECLINE,
+    });
     dispatch(playMode.acceptDrawDialog({ open: false }));
   };
