+++ src/features/mode/stockfish/dialog/EndgameSkillsDialog.js
@@ -16,7 +16,6 @@ import * as stockfishMode from 'features/mode/stockfishModeSlice';
 import * as nav from 'features/nav/navSlice';
 import multiAction from 'features/multiAction';
 import ColorButtonGroup from 'features/ColorButtonGroup';
-import Ws from 'socket/Ws';
 import styles from 'styles/dialog';
 
 const EndgameSkillsDialog = () => {
@@ -49,7 +48,13 @@ const EndgameSkillsDialog = () => {
       };
     multiAction.initGui(dispatch);
     dispatch(nav.setTraining());
-    Ws.randomizer(color, items);
+    dispatch({
+      type: 'socket/randomizer',
+      payload: {
+        color: color,
+        items: items,
+      }
+    });
   };
 
   const handleTypeChange = (event) => {
